requirements.txt
matplotlib>=3.10.8
nltk>=3.9.2
numpy>=2.4.1
pandas>=2.3.3
scikit_learn>=1.8.0

python-dotenv>=1.1.0
langchain>=1.2.7
seaborn>=0.13.2
openai>=2.14.0
langchain-openai>=1.1.9
faiss-cpu>=1.13.2